+++ src/commands/help/index.js
@@ -2,5 +2,5 @@ const help = require('./help.js');
 
 module.exports = {
     registerCommand: help.registerCommand,
-    process: help.process
+    execute: help.execute
 };