+++ src/commands/init/index.js
@@ -1,0 +1,6 @@
+const init = require('./init.js');
+
+module.exports = {
+    registerCommand: init.registerCommand,
+    execute: init.execute
+};