+++ src/resourceManager/resourceMessages.js
@@ -3,5 +3,10 @@ module.exports = {
         "usage": "Usage: noddsdale <command>",
         "errorUnknownCommand": "Unknown command: $(command)",
         "init": "Usage: noddsdale init/n noddsdale-int purpose is to prepere the current directory for a new noddsdale blog"
+    },
+    "init": {
+        "start": "Starting to initialise a Noddsdale blog",
+        "finish": "Completed initialising a Noddsdale blog",
+        "error_existing_files": "Unable to create a Noddsdale blog, the working directory contains files"
     }
 };