+++ test/noddsdale.test.js
@@ -3,7 +3,7 @@ const noddsdale = require('../src/cli'),
     assert = require('assert'),
     resources = require('../src/resourceManager/resourceMessages.js');
 
-describe('Noddsdale:', () => {
+describe('Noddsdale Integration Tests:', () => {
     let output;
 
     beforeEach(() => {
@@ -17,7 +17,7 @@ describe('Noddsdale:', () => {
     });
 
     it('Should return usage information when called with no parameters', () => {
-        noddsdale.process();
+        noddsdale.execute();
         assert.equal(output.toString(), resources.help.usage);
     });
 
@@ -26,8 +26,30 @@ describe('Noddsdale:', () => {
             errorMessage = resources.help.errorUnknownCommand,
             expectedResult = errorMessage.replace(/\$\(command\)/i, commandName);
 
-        noddsdale.process(commandName);
+        noddsdale.execute(commandName);
 
         assert.equal(output.toString(), expectedResult);
+    });
+
+    describe('Initaliise:', () => {
+        it('Should successfully initalise a new blog directory');
+
+        it.skip('Should not initalise a new blog as the working directory has existing files', () => {
+            let commandName = 'init',
+                expectedResult = resources.init.start + resources.init.error_existing_files,
+                cwd = process.cwd(),
+                testFolder = '.testrun'
+            tmp = cwd + '/' + testFolder,
+                fs = require('fs');
+            fs.mkdirSync(testFolder);
+            process.chdir(tmp);
+
+            noddsdale.execute(commandName);
+
+            assert.equal(output.toString(), expectedResult);
+
+            process.chdir(cwd);
+            fs.rmdirSync(testFolder);
+        });
     });
 });